refactor(backend-service-user): update backend-service-user to rely on backend-app-user
