+++ packages/backend/services/backend-service-user/package.json
@@ -12,13 +12,11 @@
     "@nestjs/typeorm": "9.0.1",
     "@cornie-js/api-models": "workspace:*",
     "@cornie-js/backend-api-validators": "workspace:*",
+    "@cornie-js/backend-app-jwt": "workspace:*",
+    "@cornie-js/backend-app-user": "workspace:*",
+    "@cornie-js/backend-app-user-env": "workspace:*",
     "@cornie-js/backend-common": "workspace:*",
-    "@cornie-js/backend-db": "workspace:*",
-    "@cornie-js/backend-env": "workspace:*",
     "@cornie-js/backend-http": "workspace:*",
-    "@cornie-js/backend-jwt": "workspace:*",
-    "bcrypt": "5.1.0",
-    "envalid": "7.3.1",
     "fastify": "4.17.0",
     "pg": "8.10.0",
     "reflect-metadata": "0.1.13",
@@ -27,7 +25,6 @@
   },
   "devDependencies": {
     "@jest/globals": "29.5.0",
-    "@types/bcrypt": "5.0.0",
     "@types/node": "18.16.6",
     "@typescript-eslint/eslint-plugin": "5.59.5",
     "@typescript-eslint/parser": "5.59.5",
