--- packages/backend/services/backend-service-user/src/foundation/common/adapter/nest/modules/CommonModule.ts
@@ -1,15 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { uuidProviderOutputPortSymbol } from '../../../application/ports/output/UuidProviderOutputPort';
-import { UuidProviderNodeJsAdapter } from '../../nodejs/adapters/UuidProviderNodeJsAdapter';
-
-@Module({
-  exports: [uuidProviderOutputPortSymbol],
-  providers: [
-    {
-      provide: uuidProviderOutputPortSymbol,
-      useClass: UuidProviderNodeJsAdapter,
-    },
-  ],
-})
-export class CommonModule {}
