--- packages/backend/services/backend-service-user/src/foundation/common/application/models/UuidContext.ts
@@ -1,3 +1,0 @@
-export interface UuidContext {
-  uuid: string;
-}
