--- packages/backend/services/backend-service-user/src/foundation/common/application/models/Writable.ts
@@ -1,3 +1,0 @@
-export type Writable<T> = {
-  -readonly [TKey in keyof T]: T[TKey];
-};
