--- packages/backend/services/backend-service-user/src/foundation/common/application/ports/output/UuidProviderOutputPort.ts
@@ -1,7 +1,0 @@
-export interface UuidProviderOutputPort {
-  generateV4(): string;
-}
-
-export const uuidProviderOutputPortSymbol: symbol = Symbol.for(
-  'UuidProviderOutputPort',
-);
