--- packages/backend/services/backend-service-user/src/foundation/db/adapter/nest/models/entities.ts
@@ -1,4 +1,0 @@
-import { UserDb } from '../../../../../user/adapter/typeorm/models/UserDb';
-
-// eslint-disable-next-line @typescript-eslint/ban-types
-export const typeOrmEntities: Function[] = [UserDb];
