--- packages/backend/services/backend-service-user/src/foundation/db/adapter/nest/models/migrationFolders.ts
@@ -1,3 +1,0 @@
-export const typeOrmMigrationFolders: string[] = [
-  'lib/app/adapter/typeorm/migrations/*.js',
-];
