--- packages/backend/services/backend-service-user/src/foundation/hash/adapter/nest/modules/HashModule.ts
@@ -1,15 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { bcryptHashProviderOutputPortSymbol } from '../../../application/ports/output/BcryptHashProviderOutputPort';
-import { BcryptHashProviderBcryptAdapter } from '../../bcrypt/adapters/BcryptHashProviderBcryptAdapter';
-
-@Module({
-  exports: [bcryptHashProviderOutputPortSymbol],
-  providers: [
-    {
-      provide: bcryptHashProviderOutputPortSymbol,
-      useClass: BcryptHashProviderBcryptAdapter,
-    },
-  ],
-})
-export class HashModule {}
