--- packages/backend/services/backend-service-user/src/foundation/hash/application/models/HashContext.ts
@@ -1,3 +1,0 @@
-export interface HashContext {
-  hash: string;
-}
