--- packages/backend/services/backend-service-user/src/foundation/hash/application/ports/output/BcryptHashProviderOutputPort.ts
@@ -1,8 +1,0 @@
-export interface BcryptHashProviderOutputPort {
-  hash(input: string): Promise<string>;
-  verify(input: string, hash: string): Promise<boolean>;
-}
-
-export const bcryptHashProviderOutputPortSymbol: symbol = Symbol.for(
-  'BcryptHashProviderOutputPort',
-);
