--- packages/backend/services/backend-service-user/src/user/adapter/typeorm/converters/UserCreateQueryToUserCreateQueryTypeOrmConverter.ts
@@ -1,22 +1,0 @@
-import { Converter } from '@cornie-js/backend-common';
-import { Injectable } from '@nestjs/common';
-import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity.js';
-
-import { UserCreateQuery } from '../../../domain/models/UserCreateQuery';
-import { UserDb } from '../models/UserDb';
-
-@Injectable()
-export class UserCreateQueryToUserCreateQueryTypeOrmConverter
-  implements Converter<UserCreateQuery, QueryDeepPartialEntity<UserDb>>
-{
-  public convert(
-    userCreateQuery: UserCreateQuery,
-  ): QueryDeepPartialEntity<UserDb> {
-    return {
-      email: userCreateQuery.email,
-      id: userCreateQuery.id,
-      name: userCreateQuery.name,
-      passwordHash: userCreateQuery.passwordHash,
-    };
-  }
-}
