--- packages/backend/services/backend-service-user/src/user/adapter/typeorm/converters/UserDbToUserConverter.ts
@@ -1,17 +1,0 @@
-import { Converter } from '@cornie-js/backend-common';
-import { Injectable } from '@nestjs/common';
-
-import { User } from '../../../domain/models/User';
-import { UserDb } from '../models/UserDb';
-
-@Injectable()
-export class UserDbToUserConverter implements Converter<UserDb, User> {
-  public convert(user: UserDb): User {
-    return {
-      email: user.email,
-      id: user.id,
-      name: user.name,
-      passwordHash: user.passwordHash,
-    };
-  }
-}
