--- packages/backend/services/backend-service-user/src/user/application/fixtures/UserCreateQueryV1Fixtures.ts
@@ -1,13 +1,0 @@
-import { models as apiModels } from '@cornie-js/api-models';
-
-export class UserCreateQueryV1Fixtures {
-  public static get any(): apiModels.UserCreateQueryV1 {
-    const fixture: apiModels.UserCreateQueryV1 = {
-      email: 'mail@example.com',
-      name: 'name',
-      password: 'sample input',
-    };
-
-    return fixture;
-  }
-}
