--- packages/backend/services/backend-service-user/src/user/domain/fixtures/UserCreateQueryFixtures.ts
@@ -1,15 +1,0 @@
-import { UserCreateQuery } from '../models/UserCreateQuery';
-
-export class UserCreateQueryFixtures {
-  public static get any(): UserCreateQuery {
-    const fixture: UserCreateQuery = {
-      email: 'mail@example.com',
-      id: '83073aec-b81b-4107-97f9-baa46de5dd40',
-      name: 'Name',
-      passwordHash:
-        '$2y$10$/Q/7HB2eWCzGILadcebdf.8fvya0/cnYkPdgy4q63K3IGdlnpc.7K',
-    };
-
-    return fixture;
-  }
-}
