--- packages/backend/services/backend-service-user/src/user/domain/models/UserCreateQuery.ts
@@ -1,6 +1,0 @@
-export interface UserCreateQuery {
-  readonly email: string;
-  readonly id: string;
-  readonly name: string;
-  readonly passwordHash: string;
-}
