--- packages/backend/services/backend-service-user/src/user/domain/models/UserFindQuery.ts
@@ -1,4 +1,0 @@
-export interface UserFindQuery {
-  email?: string;
-  id?: string;
-}
